Move Logging + Service Initiation out
